+++ src/App.tsx
@@ -1,31 +1,27 @@
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import { Route, Switch, Redirect } from "react-router-dom";
-import "./App.css";
-import { v4 as uuidv4 } from "uuid";
+
+import useHttp from "./hooks/use-http";
 import { useAppDispatch, useAppSelector } from "./hooks/hooks";
+import { cardsDataActions } from "./store/slice/cardsData";
+
 import Header from "./components/Header/Header";
 import Cards from "./components/Card/Cards";
-import initialData from "./state/card-data";
 import Modal from "./components/Card/CardModal/Modal";
-import useHttp from "./hooks/use-http";
 import MainPage from "./components/MainPage/MainPage";
 import TemporaryDrawer from "./components/Card/Sidebar/Sidebar";
 import CardTabs from "./components/Card/Tabs/Tabs";
 import CardDetail from "./components/Card/CardsDetail/CardsDetail";
-import { commonActions } from "./store/slice/common";
-// import { addCardActions } from "./store/slice/addCard";
-import { cardsDataActions } from "./store/slice/cardsData";
+
+import "./App.css";
 
 const App = () => {
+  const dispatch = useAppDispatch();
   const cardsData = useAppSelector((state) => state.cardsData.cards);
   const modalSelector = useAppSelector((state) => state.common.isModalActive);
   const tabSelector = useAppSelector((state) => state.tab.activeTab);
-  const dispatch = useAppDispatch();
-
-  // const state = [...initialData];
-  // const [cardList, setCardList] = useState(state);
-
   const { isLoading, error, sendRequest: fetchTasks } = useHttp();
+
   useEffect(() => {
     const transformTasks = (tasksObj: any) => {
       const loadedTasks = [];
@@ -37,23 +33,15 @@ const App = () => {
         });
       }
       dispatch(cardsDataActions.setCardsData(loadedTasks));
-      // setCardList(loadedTasks);
     };
     fetchTasks(
       { url: "https://jsonplaceholder.typicode.com/posts/" },
       transformTasks
     );
   }, [fetchTasks, dispatch]);
-  const cancelHandler = () => {
-    dispatch(commonActions.isCanceled());
-    dispatch(commonActions.setEditCardMode());
-    dispatch(commonActions.setCancelBtn());
-  };
-  const exitHandler = () => {
-    dispatch(commonActions.setIsCanceled());
-    dispatch(commonActions.setEditCardMode());
-  };
+
   let content = <p>Found no cards.</p>;
+
   if (error) content = <p className="error">{error}</p>;
   if (isLoading) content = <p>Loading...</p>;
   if (cardsData.length > 0) {
@@ -73,6 +61,7 @@ const App = () => {
       </div>
     );
   }
+
   return (
     <div className="app-wrapper">
       <Switch>
@@ -83,7 +72,7 @@ const App = () => {
           <MainPage />
         </Route>
         <Route path="/cards" exact>
-          <Header cancelHandler={cancelHandler} exitHandler={exitHandler} />
+          <Header />
           <section className="container">
             <TemporaryDrawer />
             <CardTabs />
@@ -91,7 +80,7 @@ const App = () => {
           </section>
         </Route>
         <Route exact path="/cards/:cardId">
-          <Header cancelHandler={cancelHandler} exitHandler={exitHandler} />
+          <Header />
           <section className="container">
             <TemporaryDrawer />
             <CardTabs />
@@ -106,7 +95,7 @@ const App = () => {
           </section>
         </Route>
       </Switch>
-      {modalSelector && <Modal/>}
+      {modalSelector && <Modal />}
     </div>
   );
 };
