+++ src/components/Card/Button/Button.tsx
@@ -1,18 +1,21 @@
-import React from 'react';
-
 interface ButtonProps {
-   name: string;
-   disabled: any;
-   onClick: any;
-   className: string;
+  name: string;
+  disabled: any;
+  onClick: any;
+  className: string;
 }
 
 const ButtonCard = (props: ButtonProps) => {
-  const {
-    name, disabled, onClick, className,
-  } = props;
+  const { name, disabled, onClick, className } = props;
   return (
-    <button type="button" disabled={disabled} className={className} onClick={onClick}>{name}</button>
+    <button
+      type="button"
+      disabled={disabled}
+      className={className}
+      onClick={onClick}
+    >
+      {name}
+    </button>
   );
 };
 export default ButtonCard;
