+++ src/components/MainPage/MainPage.tsx
@@ -1,4 +1,5 @@
-import React from "react";
+import { NavLink } from "react-router-dom";
+
 import Card from "@material-ui/core/Card";
 import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
@@ -7,7 +8,6 @@ import Box from "@material-ui/core/Box";
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
 import classes from "./MainPage.module.css";
-import { NavLink } from "react-router-dom";
 
 const MainPage = () => {
   return (
