+++ src/state/card-data.tsx
@@ -1,5 +1,4 @@
 import { v4 as uuidv4 } from "uuid";
-
 interface CARD_DATA {
   id: string;
   title: string;
